+++ Plugins/BidmadPlugin/Source/BidmadPlugin/Private/InterstitialCallback.cpp
@@ -1,9 +1,0 @@
-// Fill out your copyright notice in the Description page of Project Settings.
-
-
-#include "InterstitialCallback.h"
-
-FInterstitialCallback::FOnLoadAd FInterstitialCallback::OnLoadAd;
-FInterstitialCallback::FOnShowAd FInterstitialCallback::OnShowAd;
-FInterstitialCallback::FOnFailedAd FInterstitialCallback::OnFailedAd;;
-FInterstitialCallback::FOnCloseAd FInterstitialCallback::OnCloseAd;