+++ Plugins/BidmadPlugin/Source/BidmadPlugin/Private/RewardInterface.cpp
@@ -4,23 +4,58 @@
 
 #include "RewardInterface.h"
 
-TMap <FString, FRewardInterface*> FRewardInterface::mRewardInterfaceMap = TMap <FString, FRewardInterface*> ();
 
-FRewardInterface::FRewardInterface(FString zoneId)
+URewardInterface::URewardInterface()
 {
-    #if PLATFORM_ANDROID || PLATFORM_IOS
-    checkf(!zoneId.IsEmpty(), TEXT("[FRewardInterface::FRewardInterface] ZoneId is Empty #####"));
+    // Set this component to be initialized when the game starts, and to be ticked every frame.  You can turn these features
+    // off to improve performance if you don't need them.
+    PrimaryComponentTick.bCanEverTick = true;
+
+    // ...
+}
+
+// Called when the game starts
+void URewardInterface::BeginPlay()
+{
+    Super::BeginPlay();
+
+    // ...
+}
+
+
+// Called every frame
+void URewardInterface::TickComponent(float DeltaTime, ELevelTick TickType, FActorComponentTickFunction* ThisTickFunction)
+{
+    Super::TickComponent(DeltaTime, TickType, ThisTickFunction);
+
+    // ...
+}
+
+void URewardInterface::InitReward(FString androidZoneId, FString iosZoneId)
+{
+    #if PLATFORM_ANDROID
+    checkf(!androidZoneId.IsEmpty(), TEXT("[URewardInterface::InitReward] ZoneId is Empty #####"));
+    #elif PLATFORM_IOS
+    checkf(!iosZoneId.IsEmpty(), TEXT("[URewardInterface::InitReward] ZoneId is Empty #####"));
     #else
-    if(zoneId.IsEmpty()){
-        UE_LOG(FBidmadPlugin, Error, TEXT("[FRewardInterface::FRewardInterface] ZoneId is Empty #####"));
+    if(androidZoneId.IsEmpty() || androidZoneId.IsEmpty()){
+        UE_LOG(FBidmadPlugin, Error, TEXT("[URewardInterface::InitReward] ZoneId is Empty #####"));
     }
     #endif
 
-    mZoneId = zoneId;
-    GetInstance(); //Android JNI Start
+    #if PLATFORM_ANDROID
+    mZoneId = androidZoneId;
+    #else //PLATFORM_IOS 
+    mZoneId = iosZoneId;
+    #endif
+
+    //Android JNI Start
     #if PLATFORM_ANDROID && USE_ANDROID_JNI
+    GetInstance(); 
     SetActivity();
     MakeReward();
+    #elif PLATFORM_IOS
+    NewiOSInstance();
     #endif
     
     SetAdInfo(); //AOS&iOS
@@ -30,8 +65,14 @@ FRewardInterface::FRewardInterface(FString zoneId)
     #endif
 }
 
+void URewardInterface::NewiOSInstance() {
+    #if PLATFORM_ANDROID && USE_ANDROID_JNI
+    #elif PLATFORM_IOS
+    unrealReward = [[BidmadRewardInterface getSharedInstance] newInstance:[NSString stringWithUTF8String: TCHAR_TO_UTF8(*mZoneId)]];
+    #endif
+}
 
-void FRewardInterface::GetInstance() {
+void URewardInterface::GetInstance() {
     #if PLATFORM_ANDROID && USE_ANDROID_JNI
     mEnv = FAndroidApplication::GetJavaEnv();
     
@@ -43,30 +84,30 @@ void FRewardInterface::GetInstance() {
     mJObj = mEnv->CallStaticObjectMethod(mJCls, midGet, _zoneId);
     mEnv->DeleteLocalRef(_zoneId);
     #elif PLATFORM_IOS
-    unrealReward = [[BidmadRewardInterface getSharedInstance] getRewardVideoByZone:[NSString stringWithUTF8String: TCHAR_TO_UTF8(*mZoneId)]];
+    unrealReward = [[BidmadRewardInterface getSharedInstance] getInstance:[NSString stringWithUTF8String: TCHAR_TO_UTF8(*mZoneId)]];
     #endif
 }
 //Only Android Funtion START
 #if PLATFORM_ANDROID && USE_ANDROID_JNI
-void FRewardInterface::SetActivity() {
+void URewardInterface::SetActivity() {
     jmethodID midGet = FJavaWrapper::FindMethod(mEnv, mJCls, "setActivity", "(Landroid/app/Activity;)V", false);
     FJavaWrapper::CallVoidMethod(mEnv, mJObj, midGet, FAndroidApplication::GetGameActivityThis());
 }
 
-void FRewardInterface::MakeReward() {
+void URewardInterface::MakeReward() {
     jmethodID midGet = FJavaWrapper::FindMethod(mEnv, mJCls, "makeReward", "()V", false);
     mEnv->CallVoidMethod(mJObj, midGet);
     FJavaWrapper::CallVoidMethod(mEnv, mJObj, midGet);
 }
 
-void FRewardInterface::DeleteRefMember(){
+void URewardInterface::DeleteRefMember(){
     mEnv->DeleteLocalRef(mJObj);
     mEnv->DeleteGlobalRef(mJCls);
 }
 #endif
 //Only Android Funtion END
 
-void FRewardInterface::SetAdInfo() {
+void URewardInterface::SetAdInfo() {
 #if PLATFORM_ANDROID && USE_ANDROID_JNI
     jstring _zoneId = mEnv->NewStringUTF(TCHAR_TO_ANSI(*mZoneId));
     jmethodID midGet = FJavaWrapper::FindMethod(mEnv, mJCls, "setAdInfo", "(Ljava/lang/String;)V", false);
@@ -79,7 +120,7 @@ void FRewardInterface::SetAdInfo() {
     #endif
 }
 
-void FRewardInterface::Load() {
+void URewardInterface::Load() {
     GetInstance();
     #if PLATFORM_ANDROID && USE_ANDROID_JNI
     jmethodID midGet = FJavaWrapper::FindMethod(mEnv, mJCls, "load", "()V", false);
@@ -87,17 +128,11 @@ void FRewardInterface::Load() {
 
     DeleteRefMember();
     #elif PLATFORM_IOS
-    if([unrealReward isShowing]){
-        UE_LOG(LogTemp,Display, TEXT("[FRewardInterface] Current Zone Is Showing() #####"));
-        [[BidmadRewardInterface getSharedInstance] loadSameZone:[NSString stringWithUTF8String: TCHAR_TO_UTF8(*mZoneId)]];
-    }else{
-        UE_LOG(LogTemp,Display, TEXT("[FRewardInterface] Current Zone Is Not Showing() #####"));
-        [unrealReward load];
-    }
+    [unrealReward load];
     #endif
 }
 
-void FRewardInterface::Show() {
+void URewardInterface::Show() {
     GetInstance();
     #if PLATFORM_ANDROID && USE_ANDROID_JNI
     jmethodID midGet = FJavaWrapper::FindMethod(mEnv, mJCls, "show", "()V", false);
@@ -111,7 +146,7 @@ void FRewardInterface::Show() {
     #endif
 }
 
-bool FRewardInterface::IsLoaded() {
+bool URewardInterface::IsLoaded() {
     bool result = false;
     GetInstance();
     #if PLATFORM_ANDROID && USE_ANDROID_JNI
@@ -135,7 +170,13 @@ extern "C"{
         const char *zoneId = env->GetStringUTFChars(str, NULL);
         env->ReleaseStringUTFChars(str, zoneId);
 
-        FRewardCallback::OnLoadAd.Broadcast(zoneId);
+        for (TObjectIterator<URewardInterface> Itr; Itr; ++Itr)
+        {
+            if (Itr->GetWorld() != nullptr && (Itr->GetWorld()->WorldType == EWorldType::Game || Itr->GetWorld()->WorldType == EWorldType::PIE) && (!Itr->IsPendingKill()))
+            {
+                Itr->OnLoadBidmadRewardAd.Broadcast(zoneId);
+            }
+        }
     }
 
     JNIEXPORT void JNICALL Java_com_adop_sdk_reward_UnrealReward_onShowAdCb(JNIEnv *env, jobject obj, jstring str){
@@ -143,10 +184,14 @@ extern "C"{
         const char *zoneId = env->GetStringUTFChars(str, NULL);
         env->ReleaseStringUTFChars(str, zoneId);
 
-        FRewardCallback::OnShowAd.Broadcast(zoneId);
-
-        FRewardInterface ri = FRewardInterface(zoneId);
-        ri.Load();
+        for (TObjectIterator<URewardInterface> Itr; Itr; ++Itr)
+        {
+            if (Itr->GetWorld() != nullptr && (Itr->GetWorld()->WorldType == EWorldType::Game || Itr->GetWorld()->WorldType == EWorldType::PIE) && (!Itr->IsPendingKill()))
+            {
+                Itr->OnShowBidmadRewardAd.Broadcast(zoneId);
+                Itr->Load(); //Ad Reload
+            }
+        }
     }
 
     JNIEXPORT void JNICALL Java_com_adop_sdk_reward_UnrealReward_onFailedAdCb(JNIEnv *env, jobject obj, jstring str){
@@ -154,7 +199,13 @@ extern "C"{
         const char *zoneId = env->GetStringUTFChars(str, NULL);
         env->ReleaseStringUTFChars(str, zoneId);
 
-        FRewardCallback::OnFailedAd.Broadcast(zoneId);
+        for (TObjectIterator<URewardInterface> Itr; Itr; ++Itr)
+        {
+            if (Itr->GetWorld() != nullptr && (Itr->GetWorld()->WorldType == EWorldType::Game || Itr->GetWorld()->WorldType == EWorldType::PIE) && (!Itr->IsPendingKill()))
+            {
+                Itr->OnFailedBidmadRewardAd.Broadcast(zoneId);
+            }
+        }
     }
 
     JNIEXPORT void JNICALL Java_com_adop_sdk_reward_UnrealReward_onCompleteAdCb(JNIEnv *env, jobject obj, jstring str){
@@ -162,7 +213,13 @@ extern "C"{
         const char *zoneId = env->GetStringUTFChars(str, NULL);
         env->ReleaseStringUTFChars(str, zoneId);
 
-        FRewardCallback::OnCompleteAd.Broadcast(zoneId);
+        for (TObjectIterator<URewardInterface> Itr; Itr; ++Itr)
+        {
+            if (Itr->GetWorld() != nullptr && (Itr->GetWorld()->WorldType == EWorldType::Game || Itr->GetWorld()->WorldType == EWorldType::PIE) && (!Itr->IsPendingKill()))
+            {
+                Itr->OnCompleteBidmadRewardAd.Broadcast(zoneId);
+            }
+        }
     }
 
     JNIEXPORT void JNICALL Java_com_adop_sdk_reward_UnrealReward_onCloseAdCb(JNIEnv *env, jobject obj, jstring str){
@@ -170,15 +227,13 @@ extern "C"{
         const char *zoneId = env->GetStringUTFChars(str, NULL);
         env->ReleaseStringUTFChars(str, zoneId);
 
-        FRewardCallback::OnCloseAd.Broadcast(zoneId);
-    }
-
-    JNIEXPORT void JNICALL Java_com_adop_sdk_reward_UnrealReward_onClickAdCb(JNIEnv *env, jobject obj, jstring str){
-
-        const char *zoneId = env->GetStringUTFChars(str, NULL);
-        env->ReleaseStringUTFChars(str, zoneId);
-
-        FRewardCallback::OnClickAd.Broadcast(zoneId);
+        for (TObjectIterator<URewardInterface> Itr; Itr; ++Itr)
+        {
+            if (Itr->GetWorld() != nullptr && (Itr->GetWorld()->WorldType == EWorldType::Game || Itr->GetWorld()->WorldType == EWorldType::PIE) && (!Itr->IsPendingKill()))
+            {
+                Itr->OnCloseBidmadRewardAd.Broadcast(zoneId);
+            }
+        }
     }
 
     JNIEXPORT void JNICALL Java_com_adop_sdk_reward_UnrealReward_onSkippedAdCb(JNIEnv *env, jobject obj, jstring str){
@@ -186,7 +241,13 @@ extern "C"{
         const char *zoneId = env->GetStringUTFChars(str, NULL);
         env->ReleaseStringUTFChars(str, zoneId);
 
-        FRewardCallback::OnSkippedAd.Broadcast(zoneId);
+        for (TObjectIterator<URewardInterface> Itr; Itr; ++Itr)
+        {
+            if (Itr->GetWorld() != nullptr && (Itr->GetWorld()->WorldType == EWorldType::Game || Itr->GetWorld()->WorldType == EWorldType::PIE) && (!Itr->IsPendingKill()))
+            {
+                Itr->OnSkippedBidmadRewardAd.Broadcast(zoneId);
+            }
+        }
     }
 }
 #elif PLATFORM_IOS
@@ -200,75 +261,82 @@ extern "C"{
 
     return sharedInstance;
 }
-
 -(id)init{
-    if(self = [super init]){
-        NSLog(@"BidmadRewardInterface init");
-        rewardList = [[NSMutableDictionary alloc] init];
-        tempReward = nil;
-    }
-
+    self = [super init];
     return self;
 }
-
-- (UnrealReward *)getRewardVideoByZone:(NSString *)zoneID {
-    if([rewardList objectForKey:zoneID] == nil){
-        NSLog(@"NewZoneID: %@ ", zoneID);
-        UnrealReward * unrealReward = [[UnrealReward alloc] init];
-        [unrealReward setDelegate:self];
-        [rewardList setObject:unrealReward forKey:zoneID];
-    }
-    return [rewardList objectForKey:zoneID];
+- (UnrealReward *)newInstance:(NSString *)zoneID {
+    UnrealReward * unrealReward = [[UnrealReward alloc] initWithZoneId:zoneID];
+    [unrealReward setDelegate:self];
+    return unrealReward;
 }
-- (void) loadSameZone:(NSString *)zoneId{
-    tempReward = [[UnrealReward alloc] init];
-    [tempReward setZoneID:zoneId];
-    [tempReward setDelegate:self];
-    [tempReward setShowing:false];
-    [tempReward load];
-}
-- (void) updateRewardList:(NSString *) zoneId{
-    UnrealReward * temp = [rewardList objectForKey:zoneId];
-   
-    if([temp isShowing]){
-        if(tempReward != nil){
-            NSLog(@" 새로운 객체 생성  뉴페이스 등극 ");
-            [tempReward setShowing:false];
-            [rewardList setObject:tempReward forKey:zoneId];
-            tempReward = nil;
-        }
-    }
+- (UnrealReward *)getInstance:(NSString *)zoneID {
+    UnrealReward * unrealReward = [UnrealReward getInstance:zoneID];
+    return unrealReward;
 }
 - (void)BIDMADRewardSkipped:(BIDMADRewardVideo *)core {
     NSString* nsZoneId = core.zoneID;
-    FRewardCallback::OnSkippedAd.Broadcast([nsZoneId UTF8String]);
+    for (TObjectIterator<URewardInterface> Itr; Itr; ++Itr)
+    {
+        if (Itr->GetWorld() != nullptr && (Itr->GetWorld()->WorldType == EWorldType::Game || Itr->GetWorld()->WorldType == EWorldType::PIE) && (!Itr->IsPendingKill()))
+        {
+            Itr->OnSkippedBidmadRewardAd.Broadcast(nsZoneId);
+        }
+    }
 }
 - (void)BIDMADRewardVideoSucceed:(BIDMADRewardVideo *)core {
     NSString* nsZoneId = core.zoneID;
-    FRewardCallback::OnCompleteAd.Broadcast([nsZoneId UTF8String]);
+        for (TObjectIterator<URewardInterface> Itr; Itr; ++Itr)
+    {
+        if (Itr->GetWorld() != nullptr && (Itr->GetWorld()->WorldType == EWorldType::Game || Itr->GetWorld()->WorldType == EWorldType::PIE) && (!Itr->IsPendingKill()))
+        {
+            Itr->OnCompleteBidmadRewardAd.Broadcast(nsZoneId);
+        }
+    }
 }
 - (void)BIDMADRewardVideoClose:(BIDMADRewardVideo *)core {
     NSString* nsZoneId = core.zoneID;
-    [self updateRewardList:nsZoneId];
-    
-    FRewardCallback::OnCloseAd.Broadcast([nsZoneId UTF8String]);
+        for (TObjectIterator<URewardInterface> Itr; Itr; ++Itr)
+    {
+        if (Itr->GetWorld() != nullptr && (Itr->GetWorld()->WorldType == EWorldType::Game || Itr->GetWorld()->WorldType == EWorldType::PIE) && (!Itr->IsPendingKill()))
+        {
+            Itr->OnCloseBidmadRewardAd.Broadcast(nsZoneId);
+        }
+    }
 }
 - (void)BIDMADRewardVideoShow:(BIDMADRewardVideo *)core {
     NSString* nsZoneId = core.zoneID;
-    FRewardCallback::OnShowAd.Broadcast([nsZoneId UTF8String]);
-    
-    FRewardInterface ri = FRewardInterface([nsZoneId UTF8String]);
-    ri.Load();
+        for (TObjectIterator<URewardInterface> Itr; Itr; ++Itr)
+    {
+        if (Itr->GetWorld() != nullptr && (Itr->GetWorld()->WorldType == EWorldType::Game || Itr->GetWorld()->WorldType == EWorldType::PIE) && (!Itr->IsPendingKill()))
+        {
+            Itr->OnShowBidmadRewardAd.Broadcast(nsZoneId);
+            Itr->InitReward(nil, nsZoneId);
+            Itr->Load(); //Ad Reload
+        }
+    }
 }
 - (void)BIDMADRewardVideoLoad:(BIDMADRewardVideo *)core{
     NSString* nsZoneId = core.zoneID;
-    UE_LOG(LogTemp,Display, TEXT("[FRewardInterface] BIDMADRewardVideoLoad() #####"));
-    FRewardCallback::OnLoadAd.Broadcast([nsZoneId UTF8String]);
+    UE_LOG(LogTemp,Display, TEXT("[URewardInterface] BIDMADRewardVideoLoad() #####"));
+        for (TObjectIterator<URewardInterface> Itr; Itr; ++Itr)
+    {
+        if (Itr->GetWorld() != nullptr && (Itr->GetWorld()->WorldType == EWorldType::Game || Itr->GetWorld()->WorldType == EWorldType::PIE) && (!Itr->IsPendingKill()))
+        {
+            Itr->OnLoadBidmadRewardAd.Broadcast(nsZoneId);
+        }
+    }
 }
 - (void)BIDMADRewardVideoAllFail:(BIDMADRewardVideo *)core {
     NSString* nsZoneId = core.zoneID;
-    UE_LOG(LogTemp,Display, TEXT("[FRewardInterface] BIDMADRewardVideoLoad() #####"));
-    FRewardCallback::OnFailedAd.Broadcast([nsZoneId UTF8String]);
+    UE_LOG(LogTemp,Display, TEXT("[URewardInterface] BIDMADRewardVideoLoad() #####"));
+        for (TObjectIterator<URewardInterface> Itr; Itr; ++Itr)
+    {
+        if (Itr->GetWorld() != nullptr && (Itr->GetWorld()->WorldType == EWorldType::Game || Itr->GetWorld()->WorldType == EWorldType::PIE) && (!Itr->IsPendingKill()))
+        {
+            Itr->OnFailedBidmadRewardAd.Broadcast(nsZoneId);
+        }
+    }
 }
 
 @end