+++ Plugins/BidmadPlugin/Source/BidmadPlugin/Public/RewardInterface.h
@@ -3,7 +3,8 @@
 #pragma once
 
 #include "CoreMinimal.h"
-#include "RewardCallback.h"
+#include "Components/SceneComponent.h"
+
 #if PLATFORM_ANDROID
 #include "Android/AndroidApplication.h"
 
@@ -17,20 +18,31 @@
 
 #include "IOS/IOSAppDelegate.h"
 #include "IOS/IOSView.h"
-
 #include <Foundation/Foundation.h>
 #import <UIKit/UIKit.h>
-
 #pragma clang diagnostic ignored "-Wobjc-property-no-attribute"
 #import <BidmadSDK/BidmadSDK.h>
 
 #endif
 
-/**
- * 
- */
-class BIDMADPLUGIN_API FRewardInterface
+#include "RewardInterface.generated.h"
+
+DECLARE_DYNAMIC_MULTICAST_DELEGATE_OneParam(FOnLoadBidmadRewardAd, const FString&, ZoneId);
+DECLARE_DYNAMIC_MULTICAST_DELEGATE_OneParam(FOnShowBidmadRewardAd, const FString&, ZoneId);
+DECLARE_DYNAMIC_MULTICAST_DELEGATE_OneParam(FOnFailedBidmadRewardAd, const FString&, ZoneId);
+DECLARE_DYNAMIC_MULTICAST_DELEGATE_OneParam(FOnCompleteBidmadRewardAd, const FString&, ZoneId);
+DECLARE_DYNAMIC_MULTICAST_DELEGATE_OneParam(FOnOpenBidmadRewardAd, const FString&, ZoneId);
+DECLARE_DYNAMIC_MULTICAST_DELEGATE_OneParam(FOnCloseBidmadRewardAd, const FString&, ZoneId);
+DECLARE_DYNAMIC_MULTICAST_DELEGATE_OneParam(FOnClickBidmadRewardAd, const FString&, ZoneId);
+DECLARE_DYNAMIC_MULTICAST_DELEGATE_OneParam(FOnSkippedBidmadRewardAd, const FString&, ZoneId);
+
+UCLASS(Blueprintable, BlueprintType, ClassGroup=(Bidmad), meta=(BlueprintSpawnableComponent))
+class BIDMADPLUGIN_API URewardInterface : public USceneComponent   
 {
+    GENERATED_BODY()
+protected:
+    // Called when the game starts
+    virtual void BeginPlay() override;
 private:
 #if PLATFORM_ANDROID && USE_ANDROID_JNI
     void DeleteRefMember();
@@ -43,17 +55,39 @@ private:
 #elif PLATFORM_IOS
     UnrealReward* unrealReward;
 #endif
+    void NewiOSInstance();
     void GetInstance();
     void SetAdInfo();
     FString mZoneId;
 public:
-    static TMap <FString, FRewardInterface*> mRewardInterfaceMap;    
-
-    FRewardInterface(FString);
+    // Sets default values for this component's properties
+    URewardInterface();
+    // Called every frame
+    virtual void TickComponent(float DeltaTime, ELevelTick TickType, FActorComponentTickFunction* ThisTickFunction) override;  
     
+    //Bidmad Fucntion
+    UFUNCTION(BlueprintCallable,Category="BidmadReward")
+    void InitReward(FString androidZoneId, FString iosZoneId);
+    UFUNCTION(BlueprintCallable,Category="BidmadReward")
     void Load();
+    UFUNCTION(BlueprintCallable,Category="BidmadReward")
     void Show();
+    UFUNCTION(BlueprintCallable,Category="BidmadReward")
     bool IsLoaded();
+
+    //Bidmad Callback
+    UPROPERTY(BlueprintAssignable, VisibleAnywhere, BlueprintCallable, Category = "BidmadReward")
+    FOnLoadBidmadRewardAd OnLoadBidmadRewardAd;
+    UPROPERTY(BlueprintAssignable, VisibleAnywhere, BlueprintCallable, Category = "BidmadReward")
+    FOnShowBidmadRewardAd OnShowBidmadRewardAd;
+    UPROPERTY(BlueprintAssignable, VisibleAnywhere, BlueprintCallable, Category = "BidmadReward")
+    FOnFailedBidmadRewardAd OnFailedBidmadRewardAd;
+    UPROPERTY(BlueprintAssignable, VisibleAnywhere, BlueprintCallable, Category = "BidmadReward")
+    FOnCompleteBidmadRewardAd OnCompleteBidmadRewardAd;
+    UPROPERTY(BlueprintAssignable, VisibleAnywhere, BlueprintCallable, Category = "BidmadReward")
+    FOnCloseBidmadRewardAd OnCloseBidmadRewardAd;
+    UPROPERTY(BlueprintAssignable, VisibleAnywhere, BlueprintCallable, Category = "BidmadReward")
+    FOnSkippedBidmadRewardAd OnSkippedBidmadRewardAd;
 };
 
 #if PLATFORM_ANDROID && USE_ANDROID_JNI
@@ -62,20 +96,15 @@ extern "C"{
     JNIEXPORT void JNICALL Java_com_adop_sdk_reward_UnrealReward_onShowAdCb(JNIEnv *, jobject, jstring);
     JNIEXPORT void JNICALL Java_com_adop_sdk_reward_UnrealReward_onFailedAdCb(JNIEnv *, jobject, jstring);
     JNIEXPORT void JNICALL Java_com_adop_sdk_reward_UnrealReward_onCompleteAdCb(JNIEnv *, jobject, jstring);
-    JNIEXPORT void JNICALL Java_com_adop_sdk_reward_UnrealReward_onOpenAdCb(JNIEnv *, jobject, jstring);
     JNIEXPORT void JNICALL Java_com_adop_sdk_reward_UnrealReward_onCloseAdCb(JNIEnv *, jobject, jstring);
-    JNIEXPORT void JNICALL Java_com_adop_sdk_reward_UnrealReward_onClickAdCb(JNIEnv *, jobject, jstring);
     JNIEXPORT void JNICALL Java_com_adop_sdk_reward_UnrealReward_onSkippedAdCb(JNIEnv *, jobject, jstring);
 }
 #elif PLATFORM_IOS
 @interface BidmadRewardInterface : NSObject <BIDMADRewardVideoDelegate>{
-    NSMutableDictionary* rewardList;
     id<BIDMADRewardVideoDelegate> delegate;
-    UnrealReward * tempReward;
 }
 + (BidmadRewardInterface*)getSharedInstance;
-- (UnrealReward *)getRewardVideoByZone:(NSString *) zoneID;
-- (void) loadSameZone:(NSString *)zoneId;
-- (void) updateRewardList:(NSString *) zoneId;
+- (UnrealReward *)newInstance:(NSString *)zoneID;
+- (UnrealReward *)getInstance:(NSString *)zoneID;
 @end
 #endif